Fixing login test
Fixed login test so it work properly + slight modification of loginView.
